requirements.txt
a2wsgi==1.0.2
aiofiles==0.5.0
aiomysql==0.0.20
bcrypt==3.2.0
cffi==1.14.2
click==7.1.2
databases==0.3.2
h11==0.9.0
index.py==0.12.5
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.2
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
PyMySQL==0.10.0
python-multipart==0.0.5
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.8
uvicorn==0.11.8
websockets==8.1
